Move initial code to vm
