--- crates/air/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "air"
-version.workspace = true
-edition.workspace = true
-
-[dependencies]
-stwo-prover.workspace = true
-stwo-air-utils.workspace = true
-stwo-air-utils-derive.workspace = true
-rayon.workspace = true
-num-traits.workspace = true
-rand.workspace = true
-serde.workspace = true
