+++ crates/vm/Cargo.toml
@@ -1,18 +1,20 @@
 [package]
-name = "prover"
+name = "vm"
 version.workspace = true
 edition.workspace = true
 
 [dependencies]
 stwo-prover.workspace = true
+stwo-air-utils-derive.workspace = true
+stwo-air-utils.workspace = true
 tracing.workspace = true
 tracing-subscriber.workspace = true
-air.workspace = true
 num-traits.workspace = true
 thiserror.workspace = true
 serde.workspace = true
 serde_json.workspace = true
 bincode.workspace = true
+rayon.workspace = true
 
 [dev-dependencies]
 divan = "0.1.21"
