+++ crates/vm/benches/example.rs
