+++ crates/vm/src/air/components/memory.rs
@@ -27,7 +27,7 @@ use stwo_prover::{
     },
 };
 
-use crate::relations;
+use crate::air::relations;
 use serde::{Deserialize, Serialize};
 
 #[derive(Copy, Clone, Default, Serialize, Deserialize)]
