+++ crates/vm/src/air/mod.rs
@@ -1,10 +1,8 @@
-#![feature(raw_slice_split)]
-
 pub mod components;
 pub mod preprocessed;
 pub mod relations;
 
-use crate::components::{Claim, InteractionClaim};
+use components::{Claim, InteractionClaim};
 use serde::{Deserialize, Serialize};
 use stwo_prover::core::prover::StarkProof;
 use stwo_prover::core::vcs::ops::MerkleHasher;
