+++ crates/vm/src/errors.rs
