+++ crates/vm/src/lib.rs
@@ -1,0 +1,5 @@
+#![feature(raw_slice_split)]
+
+pub mod air;
+pub mod errors;
+pub mod prover;
